app/requirements/base.txt
celery>=5.2.7,<5.3
channels>=4.0.0,<5.0.0
channels_redis>=4.1.0,<5.0.0
django>=4.2,<5
django-admin-sortable2>=2.2.3,<2.3
django-cors-headers>=3.14,<3.15
django-environ>=0.10,<0.11
django-fernet>=1.0.1,<1.1
django-filter>=23.1,<23.2
django-fsm==2.7.1
django-rest-passwordreset>=1.4.0,<1.5
django-storages>=1.14,<1.15
django-userforeignkey>=0.5,<0.6
django-waffle>=4.2.0,<5.0.0
djangorestframework>=3.14,<3.15
drf-jwt>=1.19.2,<1.20
drf-spectacular[sidecar]>=0.27.2,<0.28
gunicorn[gevent]>=20.1.0,<20.2
jsonfield>=2.0.0
martor>=1.6.44,<1.7
Pillow>=9.5,<9.6
psycopg[binary]>=3.1.8,<3.2
PyJWT>=2.6,<2.7
python-dateutil>=2.9,<2.10
python-magic>=0.4,<0.5
requests>=2.28.2,<2.29
uvicorn[standard]>=0.22.0,<0.23
